error management + documentation

+++ src/Controller/SuperAdmin/SocietyController.php
@@ -15,6 +15,8 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use FOS\RestBundle\Controller\Annotations as JMS;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Swagger\Annotations as SWG;
+use Nelmio\ApiDocBundle\Annotation\Model;
 
 class SocietyController extends AbstractFOSRestController
 {
@@ -29,6 +31,22 @@ class SocietyController extends AbstractFOSRestController
      *     converter="fos_rest.request_body"
      * )
      * @IsGranted("ROLE_SUPER_ADMIN", message="Accès refusé, il faut être super admin afin d'accèder à ces informations")
+     * @SWG\Parameter(
+     *     name="Nom de la société",
+     *     in="body",
+     *     description="Ajouter une société",
+     *     @Model(type=Society::class, groups={"addsociety"})
+     * )
+     * 
+     * @SWG\Response(
+     *     response=201,
+     *     description="Ajouter une société",
+     *     @SWG\Schema(
+     *         type="array",
+     *         @SWG\Items(ref=@Model(type=Society::class))
+     *     )
+     * )
+     * @SWG\Tag(name="SuperAdmin Society")
      */
     public function addSociety(Society $society, ConstraintViolationList $violations)
     {
@@ -51,21 +69,47 @@ class SocietyController extends AbstractFOSRestController
      * )
      * @View(serializerGroups={"detail_admin_society"})
      * @IsGranted("ROLE_SUPER_ADMIN", message="Accès refusé, il faut être super admin afin d'accèder à ces informations")
+     * @SWG\Parameter(
+     *     name="id",
+     *     in="path",
+     *     type="integer",
+     *     description="L'id de la société"
+     * )
+     * @SWG\Response(
+     *     response=200,
+     *     description="Renvoie le détail de la société",
+     * )
+     * @SWG\Tag(name="SuperAdmin Society")
      */
-    public function showSociety(Society $society)
+    public function showSociety(Society $society = null, $id)
     {
-        return $society;
+        if ($society) {
+            return $society;
+        }
+
+        return $this->view("Aucune société trouvée avec l'id $id", Response::HTTP_NOT_FOUND);
     }
 
     /**
      * @Get("/api/admin/list-societies/{page}", name="app_society_list", requirements = {"page"="\d+"})
      * @JMS\View(serializerGroups={"list_society"})
      * @IsGranted("ROLE_SUPER_ADMIN", message="Accès refusé, il faut être super admin afin d'accèder à ces informations")
+     * @SWG\Parameter(
+     *     name="page",
+     *     in="path",
+     *     type="integer",
+     *     description="Choisir la page"
+     * )
+     * @SWG\Response(
+     *     response=200,
+     *     description="Renvoie la liste des sociétés",
+     * )
+     * @SWG\Tag(name="SuperAdmin Society")
      */
     public function listSocieties(PaginatorInterface $paginatorInterface, Request $request, $page)
     {
-        $societies = $paginatorInterface->paginate($this->getDoctrine()->getRepository('App\Entity\Society')->findAll(),$request->query->getInt('page',$page),5);
-        
+        $societies = $paginatorInterface->paginate($this->getDoctrine()->getRepository('App\Entity\Society')->findAll(), $request->query->getInt('page', $page), 5);
+
         return $societies->getItems();
     }
 
@@ -73,13 +117,28 @@ class SocietyController extends AbstractFOSRestController
      * @Delete("/api/admin/societies/{id}", name="app_society_delete", requirements = {"id"="\d+"})
      * @View(StatusCode = 200)
      * @IsGranted("ROLE_SUPER_ADMIN", message="Accès refusé, il faut être super admin afin d'effectuer cette action")
+     * @SWG\Parameter(
+     *     name="id",
+     *     in="path",
+     *     type="integer",
+     *     description="L'id de la société"
+     * )
+     * @SWG\Response(
+     *     response=200,
+     *     description="Suppression d'une société",
+     * )
+     * @SWG\Tag(name="SuperAdmin Society")
      */
-    public function DeleteSociety(Society $society)
+    public function DeleteSociety(Society $society = null, $id)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($society);
-        $em->flush();
+        if ($society) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($society);
+            $em->flush();
 
-        return $this->view("La société a bien été supprimée", Response::HTTP_OK);
+            return $this->view("La société a bien été supprimée", Response::HTTP_OK);
+        }
+
+        return $this->view("Aucune société trouvée avec l'id $id", Response::HTTP_NOT_FOUND);
     }
 }
